+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - **Breaking**: `aviutl2::filter::FilterPluginFlags`の`as_object`を`input`、`as_filter`を`filter`に改名
+- generic: `ObjectLayerFrame::frame_range`、`ObjectLayerFrame::frame_range_inclusive`を追加
 
 ### デモプラグイン